backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
numpy==1.26.0
python-multipart==0.0.6
scikit-learn==1.4.0
joblib==1.3.2
requests==2.31.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
python-dotenv==1.0.0
